Migrating player movement (WIP)
