+++ GateWaysToTheMultiverse/scenes/test.js
@@ -1,3 +1,5 @@
+import Player from "./assets/imports/player.js";
+
 class Test extends Phaser.Scene{
 
     constructor(){
@@ -36,32 +38,10 @@ class Test extends Phaser.Scene{
         var tiles = map.addTilesetImage("tile");
 
         //0, 0 is where we place the layer. We can use this to make rooms!!!
-        var layer = map.createStaticLayer(0, tiles, 128, 128);
+        var layer = map.createStaticLayer(0, tiles);
 
-        //THIS WILL BE MIGRATED TO A SEPARATE FILE
-        //Create the player
-        var player = this.add.sprite(160, 160, "player");
+        var player = new Player({scene:this, x:32, y:32});
 
-        //Create the keys in order to control the player
-        var key6 = this.input.keyboard.addKey("NUMPAD_SIX");
-        key6.on("down", function(event){
-            player.setPosition(player.x + 64, player.y);
-        });
-
-        var key4 = this.input.keyboard.addKey("NUMPAD_FOUR");
-        key4.on("down", function(event){
-            player.setPosition(player.x - 64, player.y);
-        });
-
-        var key2 = this.input.keyboard.addKey("NUMPAD_TWO");
-        key2.on("down", function(event){
-            player.setPosition(player.x, player.y + 64);
-        });
-
-        var key8 = this.input.keyboard.addKey("NUMPAD_EIGHT");
-        key8.on("down", function(event){
-            player.setPosition(player.x, player.y - 64);
-        });
     }
 
     update(){
